shell: add list and run endpoints
Instead of invoking in main.c, migrate list_directory and
print_directory functions to shell.c and run those using `list` and
invoke sys_run using `run`
